Add Angular front-end project
The web folder has been renamed to old_web, and a new web folder has
been created that contains an Angular project. This will be used to
store the front-end that will be created using Angular. A basic project
has been created and stored in the web folder that can be expanded upon
in the future.