Add better intregation of Angular front-end
The Angular files are now in the root of the web server, instead of
being inside a folder. This should hopefully mean that in the future
it's easier to send post requests to the Go back-end. Additionally,
routing has been added to the Angular app to allow the page to change in
response to submitting the URL form.