Add redirect to profile page after login
If the user successfully logins in or has previously logged in, they
will now be redirected to the /profile page. Despite this, there is a
bug that means a user that isn't logged in can go directly to the
profile page from the search bar and is able to access this page.